libym_logic/sat: いくつかの変更
* glueminisat-2.2.8 を追加
* YmSatMS2 を追加
* MSVC で文句を言われるところ(主にMiniSat2)の修正

+++ libraries/libym_logic/src/sat/ymsat/AssignList.h
@@ -242,7 +242,7 @@ inline
 void
 AssignList::backtrack(int level)
 {
-  ASSERT_COND( level < mSize );
+  ASSERT_COND( level < static_cast<int>(mSize) );
   mHead = mMarker[level];
   mCurLevel = level;
 }
